fix(front): load hooks on app load (#7586)
