+++ www/front_src/src/App.js
@@ -1,6 +1,7 @@
 import React, { Component } from "react";
 import Header from "./components/header";
 import { Switch, Route } from "react-router-dom";
+import { connect } from 'react-redux';
 import { ConnectedRouter } from "react-router-redux";
 import { history } from "./store";
 import ReactRouter from "./components/reactRouter";
@@ -12,6 +13,10 @@ import Footer from "./components/footer";
 import Fullscreen from 'react-fullscreen-crossbrowser';
 import queryString from 'query-string';
 import axios from './axios';
+
+import { batchActions } from "redux-batched-actions";
+import { fetchAclRoutes } from "./redux/actions/navigationActions";
+import { fetchExternalComponents } from "./redux/actions/externalComponentsActions";
 
 import styles from './App.scss';
 import footerStyles from './components/footer/footer.scss';
@@ -80,6 +85,12 @@ class App extends Component {
   }
 
   componentDidMount() {
+    const { fetchAclRoutesAndExternalComponents } = this.props;
+
+    // 1 - fetch allowed react routes
+    // 2 - fetch external components (pages, hooks...)
+    fetchAclRoutesAndExternalComponents();
+
     this.keepAlive();
   }
 
@@ -123,4 +134,13 @@ class App extends Component {
   }
 }
 
-export default App;
+const mapDispatchToProps = dispatch => {
+  return {
+    fetchAclRoutesAndExternalComponents: () => {
+      // batch actions to avoid useless multiple rendering
+      dispatch(batchActions([fetchAclRoutes(), fetchExternalComponents()]));
+    }
+  };
+};
+
+export default connect(null, mapDispatchToProps)(App);
